+++ crates/icu/src/interface/config.rs
@@ -5,5 +5,5 @@ use candid::Principal;
 pub fn get_controllers() -> Result<Vec<Principal>, Error> {
     let config = crate::config::get_config()?;
 
-    Ok(config.controllers.to_vec())
+    Ok(config.controllers)
 }