+++ crates/icu/src/interface/mod.rs
@@ -1,4 +1,5 @@
 pub mod cascade;
+pub mod config;
 pub mod ic;
 pub mod memory;
 pub mod request;