+++ crates/icu/src/interface/response.rs
@@ -21,27 +21,23 @@ pub enum Response {
 // response
 pub async fn response(req: Request) -> Result<Response, Error> {
     match req {
-        Request::CanisterCreate(cc) => create_canister(&cc.path, cc.args).await,
-        Request::CanisterUpgrade(cu) => upgrade_canister(cu.pid, &cu.path).await,
+        Request::CanisterCreate(cc) => canister_create(&cc.path, cc.extra).await,
+        Request::CanisterUpgrade(cu) => canister_upgrade(cu.pid, &cu.path).await,
     }
 }
 
-// create_canister
-async fn create_canister<A>(path: &str, args: Option<A>) -> Result<Response, Error>
+// canister_create
+async fn canister_create<A>(path: &str, extra: A) -> Result<Response, Error>
 where
     A: CandidType + Send + Sync,
 {
     let canister = canister_registry::get_canister(path)?;
     let root_pid = interface::memory::canister::state::get_root_pid()?;
+    let parent_pid = msg_caller();
+    let args = (root_pid, parent_pid, extra);
 
-    let new_canister_id = crate::interface::ic::create_canister(
-        path,
-        canister.wasm,
-        vec![root_pid],
-        msg_caller(),
-        args,
-    )
-    .await?;
+    let new_canister_id =
+        crate::interface::ic::create_canister(path, canister.wasm, vec![root_pid], args).await?;
 
     // update subnet index
     if !canister.def.is_sharded {
@@ -52,8 +48,8 @@ where
     Ok(Response::CanisterCreate(new_canister_id))
 }
 
-// upgrade_canister
-async fn upgrade_canister(pid: Principal, path: &str) -> Result<Response, Error> {
+// canister_upgrade
+async fn canister_upgrade(pid: Principal, path: &str) -> Result<Response, Error> {
     let canister = canister_registry::get_canister(path)?;
     interface::ic::upgrade_canister(pid, canister.wasm).await?;