+++ crates/icu/src/macros/start.rs
@@ -1,18 +1,13 @@
 /// icu_start
 #[macro_export]
 macro_rules! icu_start {
-    // With extra args
+    // No arg version
     (
-        $canister_path:path,
-        args = ($($arg_ty:ty),* $(,)?)
+        $canister_path:path
         $(,)?
     ) => {
         #[::icu::ic::init]
-        fn init(
-            root_pid: ::candid::Principal,
-            parent_pid: ::candid::Principal,
-            $(_arg: $arg_ty),*
-        ) {
+        fn init(root_pid: ::candid::Principal, parent_pid: ::candid::Principal) {
             use ::icu::interface::memory::canister::state;
 
             ::icu::memory::init();
@@ -21,49 +16,6 @@ macro_rules! icu_start {
             state::set_root_pid(root_pid).unwrap();
             state::set_parent_pid(parent_pid).unwrap();
             state::set_path($canister_path).unwrap();
-
-            let extra = ($(_arg),*);
-            let args = ::icu::InitArgs {
-                root_pid,
-                parent_pid,
-                extra,
-            };
-
-            _init($(_arg),*);
-        }
-
-        #[::icu::ic::update]
-        async fn init_async() {
-            _init_async().await;
-        }
-
-        ::icu::icu_endpoints!();
-    };
-
-    // Without extra args
-    (
-        $canister_path:path
-        $(,)?
-    ) => {
-        #[::icu::ic::init]
-        fn init(
-            root_pid: ::candid::Principal,
-            parent_pid: ::candid::Principal,
-        ) {
-            use ::icu::{interface::memory::canister::state};
-
-            ::icu::memory::init();
-            ::icu::log!(::icu::Log::Info, "init: {}", $canister_path);
-
-            state::set_root_pid(root_pid).unwrap();
-            state::set_parent_pid(parent_pid).unwrap();
-            state::set_path($canister_path).unwrap();
-
-            let args = ::icu::InitArgs {
-                root_pid,
-                parent_pid,
-                extra: (),
-            };
 
             _init();
         }
@@ -75,8 +27,34 @@ macro_rules! icu_start {
 
         ::icu::icu_endpoints!();
     };
-}
 
+    // Single argument version
+    (
+        $canister_path:path,
+        extra = $extra_ty:ty
+        $(,)?
+    ) => {
+        #[::icu::ic::init]
+        fn init(root_pid: ::candid::Principal, parent_pid: ::candid::Principal, extra: $extra_ty) {
+            use ::icu::interface::memory::canister::state;
+
+            ::icu::memory::init();
+            ::icu::log!(::icu::Log::Info, "init: {}", $canister_path);
+            state::set_root_pid(root_pid).unwrap();
+            state::set_parent_pid(parent_pid).unwrap();
+            state::set_path($canister_path).unwrap();
+
+            _init(extra);
+        }
+
+        #[::icu::ic::update]
+        async fn init_async() {
+            _init_async().await;
+        }
+
+        ::icu::icu_endpoints!();
+    };
+}
 /// icu_start_root
 #[macro_export]
 macro_rules! icu_start_root {