+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable, and occasionally less notable changes to this project will be docum
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.4]
+- refactored into two crates, just so I have a test crate to play with
+- updated icu_start! so it takes another optional argument to pass to the init function
+
 ## [0.1.3]
 - memory counter has now evolved into a Registry